refactor: Moved modules to own module dir
