+++ modules/doc_recognizer.py
@@ -1,6 +1,4 @@
-import tree_builder
-import util
-from exceptions import DocumentTypeRecognitionException, MissingConfigException
+from modules import util, exceptions
 
 import xml.etree.ElementTree as ET
 from io import StringIO
@@ -16,9 +14,9 @@ def recognize_doc(input_tree):
       if str(formular.text) in str(input_tree[index0][index1][index2].text):
         return formular.tag
     else:
-      raise MissingConfigException()
+      raise exceptions.MissingConfigException()
   else:
     if 1 == 1:
       return "Handbuch"
     else:
-      raise DocumentTypeRecognitionException()
+      raise exceptions.DocumentTypeRecognitionException()
